feat: Revamp how theme are loaded
All the theme code is now JS based, this will allow Saturn to be much
smoother and make the experience for adding themes much better. We are
making use of a json-like format to store themes & themesID.

This is an beta release, and currently some features are missing
including category but its in the works!